Refactor systems to handlers

+++ Components/HealthComponent.cs
@@ -2,10 +2,10 @@ namespace roguelike.Components
 {
     public class HealthComponent : Component
     {
-        public float CurrentHealth { get; set; }
-        public float MaxHealth { get; set; }
+        public double CurrentHealth { get; set; }
+        public double MaxHealth { get; set; }
 
-        public HealthComponent(float maxHealth)
+        public HealthComponent(double maxHealth)
         {
             CurrentHealth = MaxHealth = maxHealth;
         }
